v0.1.0

+++ examples/hello/hello.gsh
@@ -1,3 +1,13 @@
 mkdir "testgsh"
-ls
+lastErr!
+
+capout => {
+    ls
+}
+if lastErr != nil { // similar to: lastErr!
+    panic(lastErr)
+}
+println output.fields
+
 rmdir "testgsh"
+lastErr!
